scripts/ingestion/requirements.txt
qdrant-client>=1.9.0
openai>=1.30.0
anthropic>=0.28.0
python-dotenv>=1.0.0
tqdm>=4.66.0
tiktoken>=0.7.0
